+++ packages/create-faas-app/template/basic/src/pages/home/api/__tests__/hello.test.ts
@@ -1,13 +1,13 @@
-import { test } from '@faasjs/dev'
+import { testApi } from '@faasjs/dev'
 import { describe, it, expect } from 'vitest'
 
 import api from '../hello.api'
 
 describe('pages/home/api/hello', () => {
   it('should work', async () => {
-    const wrapped = test(api)
+    const handler = testApi(api)
 
-    const { statusCode, data } = await wrapped.JSONhandler({ name: 'world' })
+    const { statusCode, data } = await handler({ name: 'world' })
 
     expect(statusCode).toEqual(200)
     expect(data).toEqual({